file_distribution
Split code into separete files
